requirements.txt
Flask==2.2.5
python-telegram-bot==20.3
gunicorn==20.1.0
